Moved buildMultipart into BambooHTTPRequest. This seems like the most sensible place to put it, but it will B/C break for anyone who was using it raw.
