--- BambooHR/API/API.php
@@ -1,78 +1,0 @@
-<?php
-
-/*
-
-Copyright (c) 2013, Bamboo HR LLC
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted provided that the following conditions are met:
-
-* Redistributions of source code must retain the above copyright notice, this
-  list of conditions and the following disclaimer.
-
-* Redistributions in binary form must reproduce the above copyright notice,
-  this list of conditions and the following disclaimer in the documentation
-  and/or other materials provided with the distribution.
-
-* Neither the name of Bamboo HR nor the names of its contributors may be used
-  to endorse or promote products derived from this software without specific
-  prior written permission.
-
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-*/
-namespace BambooHR\API;
-
-define('BAMBOOHR_MULTIPART_BOUNDARY', "----BambooHR-MultiPart-Mime-Boundary----");
-
-/**
- * Utility function for creating a valid multipart POST body. Mostly useful for uploading files.
- *
- * @param type $boundary
- * @param type $postFields
- * @param type $name
- * @param type $fileName
- * @param type $contentType
- * @param type $fileData
- * @return string
- */
-function buildMultipart($boundary, $postFields, $name, $fileName, $contentType, $fileData ) {
-
-	$data = '';
-
-	// populate normal fields first (simpler)
-	foreach ($postFields as $key => $content) {
-	    $data .= "--".$boundary. "\r\n";
-	    $data .= 'Content-Disposition: form-data; name="' . $key . '"';
-	    // note: double endline
-	    $data .= "\r\n\r\n";
-	    $data.= $content."\r\n";
-	}
-	$data.="--".$boundary."\r\n";
-    	$data .= 'Content-Disposition: form-data; name="' . $name . '";' .
-             ' filename="' . $fileName . '"' . "\r\n";
-
-   	$data .= 'Content-Type: ' . $contentType . "\r\n";
-    	$data .= "\r\n";
-    	// the file itself (note: there's no encoding of any kind)
-
-    	$data .= $fileData . "\r\n";
-	// last delimiter
-	$data .= "--" . $boundary . "--\r\n";
-	return $data;
-}
-
-
-
-?>
