+++ BambooHR/BambooAPI.php
@@ -546,8 +546,8 @@ class BambooAPI {
             "fileName"=>$fileName,
             "share"=>$shareWithEmployees
         );
-        $request->content=buildMultipart( BAMBOOHR_MULTIPART_BOUNDARY, $params, "file",$fileName,$contentType, $fileData);
-        $request->headers[]="Content-Type: multipart/form-data; boundary=".BAMBOOHR_MULTIPART_BOUNDARY;
+        $request->content=BambooHTTPRequest::buildMultipart( BambooHTTPRequest::BAMBOOHR_MULTIPART_BOUNDARY, $params, "file",$fileName,$contentType, $fileData);
+        $request->headers[]="Content-Type: multipart/form-data; boundary=".BambooHTTPRequest::BAMBOOHR_MULTIPART_BOUNDARY;
         $request->headers[]="Content-Length: ".strlen( $request->content );
 
         return $this->httpHandler->sendRequest( $request );
@@ -590,8 +590,8 @@ class BambooAPI {
             "fileName"=>$fileName,
             "share"=>$shareWithEmployees
         );
-        $request->content=buildMultipart( BAMBOOHR_MULTIPART_BOUNDARY, $params, "file",$fileName,$contentType, $fileData);
-        $request->headers[]="Content-Type: multipart/form-data; boundary=".BAMBOOHR_MULTIPART_BOUNDARY;
+        $request->content=BambooHTTPRequest::buildMultipart( BambooHTTPRequest::BAMBOOHR_MULTIPART_BOUNDARY, $params, "file",$fileName,$contentType, $fileData);
+        $request->headers[]="Content-Type: multipart/form-data; boundary=".BambooHTTPRequest::BAMBOOHR_MULTIPART_BOUNDARY;
         $request->headers[]="Content-Length: ".strlen( $request->content );
 
         return $this->httpHandler->sendRequest( $request );
